Always update 'All tracks' playlist on rescan
Tracks that are explicitly removed by the user won't be added back, a string set preference is used to store the track ids deleted from 'All tracks' playlist. Of course, there are other ways to do this (with or without db modification) but using a preference turned out to be the simplest solution.