Do player operations on a background thread
As per https://developer.android.com/guide/topics/media/exoplayer/hello-world#a-note-on-threading, using the main thread is suitable for majority of cases but it still slows down the main thread, in some cases significantly. This can lead to a non-responsive UI when doing player operations like `moveMediaItems()` with a lot of media items in the queue.